Move stuff from unit class to movable
Movable has everything related to units moving. Also split limits into new file (vs_limits.h). Can't use limits.h, as that's a system header. Finally, split computer into a separate file.
