+++ engine/src/cmd/ai/aggressive.cpp
@@ -287,7 +287,7 @@ bool AggressiveAI::ProcessLogicItem( const AIEvents::AIEvresult &item )
                 value = ( pdmag-parent->rSize()-targ->rSize() );
                 float  myvel = PosDifference.Dot( parent->GetVelocity()-targ->GetVelocity() )/value;        ///pdmag;
                 if (myvel > 0)
-                    value -= myvel*myvel/( 2*( parent->Limits().retro/parent->GetMass() ) );
+                    value -= myvel*myvel/( 2*( parent->limits.retro/parent->GetMass() ) );
             } else {
                 value = 10000;
             }
